restructure omnia crate
