+++ crates/omnia/src/deployment.rs
@@ -1,5 +1,10 @@
 //! # WebAssembly Initiator
 
+mod manifest;
+mod source;
+
+pub use source::{LoadedGuest, Source};
+
 use std::collections::BTreeSet;
 use std::env;
 use std::path::{Path, PathBuf};
@@ -11,12 +16,10 @@ use wasmtime::{Config, Engine};
 use wasmtime_wasi::WasiView;
 use wrpc_wasmtime::WrpcView;
 
-use crate::dispatch::DispatchHandle;
-use crate::manifest::Manifest;
-use crate::registry::{Registry, RegistryBuilder};
-use crate::routing::Routes;
-use crate::selector::{FirstArgSelector, GuestSelector};
-use crate::source::{LoadedGuest, Source};
+use crate::dispatch::{DispatchHandle, FirstArgSelector, GuestSelector};
+use crate::registry::{Registry, RegistryBuilder, Routes};
+
+use manifest::Manifest;
 use crate::working_tree::{ResolvedPreopen, WorkingTreeRegistry};
 use crate::{Host, RuntimeOptions, Telemetry};
