+++ crates/omnia/src/dispatch.rs
@@ -25,586 +25,16 @@
 //! and instance-per-call serve integration ([`ServeExt::serve_function`]) for
 //! the carrier round-trip.
 
-use std::collections::{BTreeSet, HashMap};
-use std::iter::zip;
-use std::pin::pin;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::{Arc, OnceLock};
+mod handle;
+mod host;
+mod link;
+mod selector;
+mod serve;
+mod transport;
 
-use anyhow::{Context as _, Result, bail, ensure};
-use bytes::BytesMut;
-use futures::{FutureExt as _, StreamExt as _};
-use tokio_util::codec::Encoder as _;
-use wasmtime::component::{Linker, Type, Val, types};
-use wasmtime::{AsContextMut as _, Engine, StoreContextMut};
-use wasmtime_wasi::WasiView;
-use wrpc_transport::Invoke;
-use wrpc_wasmtime::{ServeExt as _, ValEncoder, WrpcView, read_value};
-
-use crate::registry::GuestId;
-use crate::runtime::Runtime;
-use crate::selector::GuestSelector;
-use crate::source::LoadedGuest;
-use crate::traits::FutureResult;
-use crate::transport::{InProcClient, InProcServer, InProcess, LinkTransport as _};
-
-/// wRPC host-resource map shape (empty for the resource-free dynamic path).
-type HostResources = HashMap<
-    Box<str>,
-    HashMap<Box<str>, (wasmtime::component::ResourceType, wasmtime::component::ResourceType)>,
->;
-
-/// Wire the serve side of every host-mediated interface.
-///
-/// Each target guest that exports a linked interface runs a wRPC server whose
-/// handlers instantiate the guest *fresh per call* (instance-per-call); the
-/// bound transport is then installed so polyfilled imports can reach it.
-///
-/// Spawns one detached task per served function to drain its invocation stream.
-/// A no-op when no guest declares any `link` interface.
-///
-/// # Errors
-///
-/// Returns an error if a guest's export cannot be served over the carrier.
-pub async fn serve_links<B>(state: &Runtime<B>) -> Result<()>
-where
-    B: Clone + Send + Sync + 'static,
-{
-    let registry = state.registry();
-    let handle = registry.dispatch();
-    if handle.links().is_empty() {
-        return Ok(());
-    }
-    let engine = registry.engine().clone();
-
-    let mut servers: HashMap<GuestId, Arc<InProcServer>> = HashMap::new();
-    for guest in registry.guests() {
-        let component_ty = guest.component().component_type();
-        let mut server: Option<Arc<InProcServer>> = None;
-
-        for (interface, types::ComponentExtern { ty, .. }) in component_ty.exports(&engine) {
-            if !handle.links().contains(interface) {
-                continue;
-            }
-            let types::ComponentItem::ComponentInstance(instance_ty) = ty else {
-                continue;
-            };
-            for (func, types::ComponentExtern { ty, .. }) in instance_ty.exports(&engine) {
-                let types::ComponentItem::ComponentFunc(func_ty) = ty else {
-                    continue;
-                };
-                let server =
-                    Arc::clone(server.get_or_insert_with(|| Arc::new(InProcServer::default())));
-                let runtime = state.clone();
-                let factory = move || runtime.build_store(runtime.store());
-                let stream = server
-                    .serve_function(
-                        factory,
-                        guest.instance_pre().clone(),
-                        Arc::<HostResources>::default(),
-                        func_ty,
-                        interface,
-                        func,
-                    )
-                    .await
-                    .with_context(|| {
-                        format!("serving `{interface}/{func}` from guest `{}`", guest.id())
-                    })?;
-
-                tokio::spawn(async move {
-                    let mut stream = pin!(stream);
-                    while let Some(invocation) = stream.next().await {
-                        match invocation {
-                            Ok((_cx, fut)) => {
-                                tokio::spawn(async move {
-                                    if let Err(error) = fut.await {
-                                        tracing::error!(%error, "link serve invocation failed");
-                                    }
-                                });
-                            }
-                            Err(error) => tracing::error!(%error, "link serve accept failed"),
-                        }
-                    }
-                });
-            }
-        }
-
-        if let Some(server) = server {
-            servers.insert(guest.id().clone(), server);
-        }
-    }
-
-    handle.install(InProcess::new(servers));
-    Ok(())
-}
-
-/// Host-originated dynamic dispatch into a *known* guest export — the host→guest
-/// counterpart of the selector-driven guest→guest `dispatch`.
-///
-/// Shares the depth bound (`DispatchHandle::enter`) and resource rejection with
-/// guest→guest dispatch. The target is instantiated *fresh* on a new store and
-/// the matching export invoked directly, so the callee can never re-enter its
-/// caller and needs no `link` declaration for `interface`.
-///
-/// `args` and the returned values are plain `Val`s; a live resource handle on
-/// either side is rejected.
-///
-/// # Errors
-///
-/// Returns an error if the depth bound is exceeded, an argument or result carries
-/// a resource handle, the target is not registered, the named `interface`/`func`
-/// export is absent or is not a function, or the guest call traps.
-pub async fn dispatch<B>(
-    runtime: &Runtime<B>, target: &GuestId, interface: &str, func: &str, args: Vec<Val>,
-) -> Result<Vec<Val>>
-where
-    B: Clone + Send + Sync + 'static,
-{
-    // Depth-count this hop exactly like a guest→guest dispatch. The guard
-    // is held here (borrowing `runtime`) across the awaited callee task below.
-    let _guard = runtime.registry().dispatch().enter(target)?;
-
-    // Plain records cross by value; a live resource handle never crosses.
-    for value in &args {
-        if contains_resource(value) {
-            bail!(
-                "a resource handle cannot cross the link seam \
-                 (host→guest `{interface}/{func}`, target `{target}`)"
-            );
-        }
-    }
-
-    let instance_pre = runtime
-        .registry()
-        .get(target)
-        .with_context(|| format!("dispatch target `{target}` is not registered"))?
-        .instance_pre()
-        .clone();
-
-    // Run the callee on its own task. `resolve` is invoked from *within* the
-    // caller guest's concurrent event loop (the backend's loop awaits it inside
-    // the `complete` host call), and wasmtime forbids a recursive
-    // `StoreContextMut::run_concurrent` on the same thread. Spawning gives the
-    // callee a fresh event loop: when the caller's loop parks awaiting this task,
-    // its ambient store clears, so the callee's call runs unnested. The task owns
-    // the whole store lifecycle (build → instantiate → call → drop), so the
-    // callee is a fresh instance that cannot re-enter its caller
-    // (instance-per-call) and needs no `link` declaration for `interface`.
-    let task_runtime = (*runtime).clone();
-    let target_owned = target.clone();
-    let interface_owned = interface.to_owned();
-    let func_owned = func.to_owned();
-    let results = tokio::spawn(async move {
-        let mut store = task_runtime.build_store(task_runtime.store());
-        let instance = task_runtime
-            .instantiate(&instance_pre, &mut store)
-            .await
-            .with_context(|| format!("instantiating dispatch target `{target_owned}`"))?;
-
-        let interface_idx =
-            instance.get_export_index(&mut store, None, &interface_owned).with_context(|| {
-                format!("guest `{target_owned}` exports no interface `{interface_owned}`")
-            })?;
-        let (item, func_idx) = instance
-            .get_export(&mut store, Some(&interface_idx), &func_owned)
-            .with_context(|| {
-            format!(
-                "interface `{interface_owned}` (guest `{target_owned}`) exports no \
-                     `{func_owned}`"
-            )
-        })?;
-        let types::ComponentItem::ComponentFunc(func_ty) = item else {
-            bail!("`{interface_owned}/{func_owned}` (guest `{target_owned}`) is not a function");
-        };
-        let result_count = func_ty.results().count();
-        let function = instance.get_func(&mut store, func_idx).with_context(|| {
-            format!("resolving `{interface_owned}/{func_owned}` on guest `{target_owned}`")
-        })?;
-
-        let mut results = vec![Val::Bool(false); result_count];
-        function
-            .call_async(&mut store, &args, &mut results)
-            .await
-            .map_err(anyhow::Error::from)
-            .with_context(|| {
-                format!("calling `{interface_owned}/{func_owned}` on guest `{target_owned}`")
-            })?;
-        Ok::<Vec<Val>, anyhow::Error>(results)
-    })
-    .await
-    .with_context(|| format!("joining dispatch target `{target}` task"))?
-    .with_context(|| format!("dispatching `{interface}/{func}` to guest `{target}`"))?;
-
-    // A target must not hand back a resource handle either.
-    for value in &results {
-        if contains_resource(value) {
-            bail!(
-                "a resource handle cannot cross the link seam \
-                 (result of `{interface}/{func}`, target `{target}`)"
-            );
-        }
-    }
-
-    Ok(results)
-}
-
-/// A host→guest call capability, type-erased so a host binding (e.g.
-/// `wasi-model`'s `resolve`) can invoke a guest without naming the concrete
-/// [`Runtime`].
-///
-/// The `runtime!` macro threads an `Arc<dyn HostDispatch>` into each store
-/// context so any host binding gets dynamic host→guest calls for free. It
-/// carries no consumer vocabulary — a consumer owns its own verb names and
-/// return shapes and composes this generic seam.
-pub trait HostDispatch: Send + Sync + 'static {
-    /// Invoke `target`'s `interface`/`func` with `args`, returning the typed
-    /// results. The target is instantiated *fresh* (instance-per-call), the hop
-    /// is depth-bounded like any host-mediated call, and a live resource handle
-    /// on either side is rejected.
-    ///
-    /// A `None` `interface` discovers the unique exported interface carrying a
-    /// function named `func` — a structural component-model query that names no
-    /// consumer scheme.
-    fn invoke(
-        &self, target: GuestId, interface: Option<String>, func: String, args: Vec<Val>,
-    ) -> FutureResult<Vec<Val>>;
-}
-
-impl<B: Clone + Send + Sync + 'static> HostDispatch for Runtime<B> {
-    fn invoke(
-        &self, target: GuestId, interface: Option<String>, func: String, args: Vec<Val>,
-    ) -> FutureResult<Vec<Val>> {
-        let runtime = self.clone();
-        async move {
-            let interface: Box<str> = match interface {
-                Some(name) => Box::from(name),
-                None => find_interface(&runtime, &target, &func)?,
-            };
-            dispatch(&runtime, &target, &interface, &func, args).await
-        }
-        .boxed()
-    }
-}
-
-/// Find the exported interface on `target`'s component that carries a function
-/// named `func`, so a host can invoke it without hardcoding a consumer
-/// interface name.
-fn find_interface<B: Clone + Send + Sync + 'static>(
-    runtime: &Runtime<B>, target: &GuestId, func: &str,
-) -> Result<Box<str>> {
-    let registry = runtime.registry();
-    let engine = registry.engine();
-    let guest = registry
-        .get(target)
-        .with_context(|| format!("dispatch target `{target}` is not registered"))?;
-    let component_ty = guest.component().component_type();
-    for (interface, types::ComponentExtern { ty, .. }) in component_ty.exports(engine) {
-        let types::ComponentItem::ComponentInstance(instance_ty) = ty else {
-            continue;
-        };
-        let has_func =
-            instance_ty.exports(engine).any(|(name, types::ComponentExtern { ty, .. })| {
-                name == func && matches!(ty, types::ComponentItem::ComponentFunc(_))
-            });
-        if has_func {
-            return Ok(Box::from(interface));
-        }
-    }
-    bail!("dispatch target `{target}` exports no interface with a `{func}` function")
-}
-
-/// The long-lived dispatch state shared by every polyfilled import.
-///
-/// It carries the selector strategy, the union of host-mediated interfaces, the
-/// bound transport (installed once the serve side is wired), and the
-/// process-wide dispatch-depth counter.
-pub struct DispatchHandle {
-    selector: Arc<dyn GuestSelector>,
-    links: BTreeSet<Box<str>>,
-    transport: OnceLock<InProcess>,
-    depth: AtomicUsize,
-    max_depth: usize,
-}
-
-impl DispatchHandle {
-    /// Create a shared dispatch handle. The transport is installed later by
-    /// [`serve_links`], once each target's serve side is wired.
-    #[must_use]
-    pub fn new(
-        selector: Arc<dyn GuestSelector>, links: BTreeSet<Box<str>>, max_depth: usize,
-    ) -> Arc<Self> {
-        Arc::new(Self {
-            selector,
-            links,
-            transport: OnceLock::new(),
-            depth: AtomicUsize::new(0),
-            max_depth,
-        })
-    }
-
-    /// The union of host-mediated interface names across every guest's `link`
-    /// allow-list — the set of interfaces to polyfill (caller side) and serve
-    /// (callee side).
-    #[must_use]
-    pub const fn links(&self) -> &BTreeSet<Box<str>> {
-        &self.links
-    }
-
-    /// Install the bound transport carrier (called once by [`serve_links`]).
-    fn install(&self, transport: InProcess) {
-        // Set-once: a second install (there is only ever one) is ignored.
-        let _ = self.transport.set(transport);
-    }
-
-    /// The bound transport carrier.
-    fn transport(&self) -> Result<&InProcess> {
-        self.transport
-            .get()
-            .context("link transport not initialized; `serve_links` must run before dispatch")
-    }
-
-    /// Enter a dispatch, bounding nesting depth. The returned guard decrements
-    /// the shared counter on drop.
-    ///
-    /// The counter is process-wide and tracks *synchronous* nesting (A->B->C,
-    /// each awaited to completion before the caller returns); it is a safety
-    /// bound, not a precise per-chain limit under heavy concurrency.
-    fn enter(&self, target: &GuestId) -> Result<DepthGuard<'_>> {
-        let depth = self.depth.fetch_add(1, Ordering::SeqCst) + 1;
-        if depth > self.max_depth {
-            self.depth.fetch_sub(1, Ordering::SeqCst);
-            bail!(
-                "link dispatch depth {depth} exceeds maximum {} (target `{target}`); raise \
-                 MAX_DISPATCH_DEPTH if this is intentional",
-                self.max_depth
-            );
-        }
-        Ok(DepthGuard { depth: &self.depth })
-    }
-}
-
-/// Decrements the shared dispatch-depth counter when a dispatch unwinds.
-#[derive(Debug)]
-struct DepthGuard<'a> {
-    depth: &'a AtomicUsize,
-}
-
-impl Drop for DepthGuard<'_> {
-    fn drop(&mut self) {
-        self.depth.fetch_sub(1, Ordering::SeqCst);
-    }
-}
-
-/// Polyfill every host-mediated import named in the `link` allow-list union onto
-/// the shared linker, bound to the dispatch handle.
-///
-/// Each interface is linked exactly once (the linker is shared, so the per-guest
-/// allow-lists are unioned). `wasi:*` imports are never touched here — they are
-/// host-satisfied — so only the manifest-declared interfaces are dispatched.
-///
-/// Runs *before* pre-instantiation, so an import that is neither host-satisfied
-/// nor allow-listed remains unresolved and fails fast at `instantiate_pre`.
-///
-/// # Errors
-///
-/// Returns an error if a named link target is not an interface import, or if a
-/// function cannot be defined on the linker.
-pub fn link<T>(
-    engine: &Engine, linker: &mut Linker<T>, guests: &[LoadedGuest], handle: &Arc<DispatchHandle>,
-) -> Result<()>
-where
-    T: WasiView + WrpcView + 'static,
-{
-    if handle.links.is_empty() {
-        return Ok(());
-    }
-
-    let mut wired: BTreeSet<Box<str>> = BTreeSet::new();
-    for LoadedGuest { id, component } in guests {
-        let component_ty = component.component_type();
-        for (name, types::ComponentExtern { ty, .. }) in component_ty.imports(engine) {
-            if !handle.links.contains(name) || wired.contains(name) {
-                continue;
-            }
-            let types::ComponentItem::ComponentInstance(instance_ty) = ty else {
-                bail!("link target `{name}` (imported by guest `{id}`) is not an interface");
-            };
-
-            // Snapshot the interface's function types before mutably borrowing
-            // the linker.
-            let funcs: Vec<Arc<str>> = instance_ty
-                .exports(engine)
-                .filter_map(|(func, types::ComponentExtern { ty, .. })| {
-                    matches!(ty, types::ComponentItem::ComponentFunc(_)).then(|| Arc::from(func))
-                })
-                .collect();
-
-            let mut root = linker.root();
-            let mut interface = root
-                .instance(name)
-                .map_err(anyhow::Error::from)
-                .with_context(|| format!("defining host-mediated interface `{name}`"))?;
-            let iface_name: Arc<str> = Arc::from(name);
-
-            for func in &funcs {
-                let handle = Arc::clone(handle);
-                let iface_name = Arc::clone(&iface_name);
-                let func_name = Arc::clone(func);
-                interface
-                    .func_new_async(func, move |store, ty, params, results| {
-                        let handle = Arc::clone(&handle);
-                        let iface_name = Arc::clone(&iface_name);
-                        let func_name = Arc::clone(&func_name);
-                        Box::new(async move {
-                            send(store, &handle, &iface_name, &func_name, &ty, params, results)
-                                .await
-                                .map_err(wasmtime::Error::from_anyhow)
-                        })
-                    })
-                    .map_err(anyhow::Error::from)
-                    .with_context(|| format!("polyfilling `{name}` function `{func}`"))?;
-            }
-            wired.insert(Box::from(name));
-        }
-    }
-    Ok(())
-}
-
-/// The per-call dispatch: select the target, reject crossing resources, bound
-/// depth, then round-trip the call over the in-process wRPC carrier to a
-/// freshly-instantiated target export.
-async fn send<T>(
-    mut store: StoreContextMut<'_, T>, handle: &DispatchHandle, interface: &str, func: &str,
-    ty: &types::ComponentFunc, params: &[Val], results: &mut [Val],
-) -> Result<()>
-where
-    T: WrpcView + 'static,
-{
-    let start = std::time::Instant::now();
-
-    let (target, forwarded) = handle
-        .selector
-        .select(interface, func, params)
-        .with_context(|| format!("selecting target for `{interface}/{func}`"))?;
-
-    // Plain records cross by value; a live resource handle never crosses.
-    for value in &forwarded {
-        if contains_resource(value) {
-            bail!(
-                "a resource handle cannot cross the link seam (call to `{interface}/{func}`, \
-                 target `{target}`)"
-            );
-        }
-    }
-
-    let _guard = handle.enter(&target)?;
-
-    let param_types: Vec<Type> = ty.params().map(|(_, ty)| ty).collect();
-    let result_types: Vec<Type> = ty.results().collect();
-    ensure!(
-        forwarded.len() == param_types.len(),
-        "selector forwarded {} arguments but `{interface}/{func}` expects {}",
-        forwarded.len(),
-        param_types.len()
-    );
-
-    let client = handle.transport()?.connect(&target)?;
-
-    // Encode the forwarded parameters with wRPC's value codec.
-    let mut buf = BytesMut::new();
-    for (value, ty) in zip(&forwarded, &param_types) {
-        let mut encoder: ValEncoder<'_, T, <InProcClient as Invoke>::Outgoing> =
-            ValEncoder::new(store.as_context_mut(), ty, &[], &[]);
-        encoder
-            .encode(value, &mut buf)
-            .map_err(anyhow::Error::from)
-            .with_context(|| format!("encoding parameter for `{interface}/{func}`"))?;
-        ensure!(
-            encoder.deferred.is_none(),
-            "async/stream parameters cannot cross the link seam (`{interface}/{func}`)"
-        );
-    }
-
-    // Invoke over the carrier; the request is written and flushed here, the
-    // results stream back on `incoming`. No deferred (async) parameters, so the
-    // outgoing half carries nothing further and is dropped.
-    let (_outgoing, incoming) = client
-        .invoke((), interface, func, buf.freeze(), &[[]; 0])
-        .await
-        .with_context(|| format!("invoking link target `{target}` for `{interface}/{func}`"))?;
-
-    let mut incoming = pin!(incoming);
-    for (index, (value, ty)) in zip(results.iter_mut(), &result_types).enumerate() {
-        read_value(&mut store, &mut incoming, &[], &[], value, ty, &[index])
-            .await
-            .map_err(anyhow::Error::from)
-            .with_context(|| format!("decoding result {index} from `{target}`"))?;
-    }
-
-    let elapsed_us = u64::try_from(start.elapsed().as_micros()).unwrap_or(u64::MAX);
-    tracing::debug!(
-        target = %target,
-        interface,
-        func,
-        transport = "in-process",
-        histogram.link_dispatch_duration_us = elapsed_us,
-        monotonic_counter.link_dispatches = 1_u64,
-        "dispatched host-mediated call",
-    );
-    Ok(())
-}
-
-/// Recursively reports whether a value carries a live resource handle.
-fn contains_resource(value: &Val) -> bool {
-    match value {
-        Val::Resource(_) => true,
-        Val::List(values) | Val::Tuple(values) => values.iter().any(contains_resource),
-        Val::Record(fields) => fields.iter().any(|(_, value)| contains_resource(value)),
-        Val::Variant(_, Some(value))
-        | Val::Option(Some(value))
-        | Val::Result(Ok(Some(value)) | Err(Some(value))) => contains_resource(value),
-        _ => false,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-    use std::sync::atomic::Ordering;
-
-    use wasmtime::component::Val;
-
-    use super::{DispatchHandle, contains_resource};
-    use crate::registry::GuestId;
-    use crate::selector::FirstArgSelector;
-
-    fn handle(max_depth: usize) -> Arc<DispatchHandle> {
-        DispatchHandle::new(Arc::new(FirstArgSelector), std::iter::empty().collect(), max_depth)
-    }
-
-    #[test]
-    fn depth_guard() {
-        let handle = handle(2);
-        let target = GuestId::from("t");
-
-        let first = handle.enter(&target).expect("depth 1 within bound");
-        let second = handle.enter(&target).expect("depth 2 within bound");
-        handle.enter(&target).expect_err("depth 3 exceeds the maximum");
-
-        // Unwinding the guards frees the budget again.
-        drop(second);
-        drop(first);
-        assert_eq!(handle.depth.load(Ordering::SeqCst), 0);
-        handle.enter(&target).expect("budget freed after guards drop");
-    }
-
-    #[test]
-    fn detect_nested() {
-        // Plain values never count as resources.
-        assert!(!contains_resource(&Val::String("x".to_owned())));
-        assert!(!contains_resource(&Val::Record(vec![("f".to_owned(), Val::U32(1),)])));
-        assert!(!contains_resource(&Val::Option(None)));
-        // A nested option/list carrying plain values stays clean.
-        assert!(!contains_resource(&Val::List(vec![Val::Bool(true), Val::Bool(false)])));
-    }
-}
+pub use handle::DispatchHandle;
+pub use host::HostDispatch;
+pub use link::link;
+pub use selector::{FirstArgSelector, GuestSelector};
+pub use serve::serve_links;
+pub use transport::{LinkClient, WrpcState};
